% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_functions.R
\name{ApproximationBasedCelltypeAssignment}
\alias{ApproximationBasedCelltypeAssignment}
\title{ELeFHAnt Approximation based Celltype Annotation}
\usage{
ApproximationBasedCelltypeAssignment(
  reference = reference,
  query = query,
  downsample = downsample,
  downsample_to = downsample_to,
  classification.method = classification.method,
  crossvalidationSVM = crossvalidationSVM,
  validatePredictions = validatePredictions,
  selectvarfeatures = selectvarfeatures,
  ntree = ntree
)
}
\arguments{
\item{reference}{a processed Seurat Object with Celltypes column in metadata}

\item{query}{a processed seurat object with seurat_clusters column in metadata}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample reference and query enabling fast computation}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 100} in reference and query for Celltypes and seurat_clusters resspectively}

\item{classification.method}{choose classification method for learning and predicting celltypes.
Choices: randomForest (decision trees), SVM (Support Vector Machines) or Ensemble (uses estimation robustness of both randomForest and SVM to predict)}

\item{crossvalidationSVM}{if a integer value k>0 is specified, a k-fold cross validation on the training data is performed to assess the quality of the model}

\item{validatePredictions}{logical indicator (TRUE or FALSE) to asses predictions by deploying Gene set enrichment analysis}

\item{selectvarfeatures}{number of variable features to select while training (default: 2000)}

\item{ntree}{number of trees randomForest classifier should build (Default: 500)}
}
\value{
query seurat object with predictions added to meta.data of the object
}
\description{
Approximation based Celltype Annotation is a function to annotate cells in a single cell datasets using approximation. Approximation sets highest voted cell type for a query cluster.
It requires a reference dataset (a processed Seurat Object with Celltypes column in metadata) and a query dataset (a processed
seurat object with seurat_clusters column in metadata). One can choose from randomForest, SVM or Ensemble classifiction method
to learn celltypes from reference dataset and then predict celltypes for query dataset.
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
