% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_Functions.R
\name{BenchmarkELeFHAnt}
\alias{BenchmarkELeFHAnt}
\title{Benchmark ELeFHAnt against scPred and Label Transfer}
\usage{
BenchmarkELeFHAnt(
  reference = reference,
  query = query,
  downsample = TRUE,
  downsample_to = 1000,
  selectvarfeatures = 2000,
  ntree = 500,
  annotationCol = "Celltypes"
)
}
\arguments{
\item{reference}{a processed Seurat Object with Celltypes column in metadata}

\item{query}{a processed seurat object with seurat_clusters column in metadata}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample reference, enabling fast computation.}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 1000}}

\item{selectvarfeatures}{number of variable features to select while training (Default: 2000)}

\item{ntree}{number of trees randomForest classifier should build (Default: 500)}

\item{annotationCol}{Name of column in metdata which has Celltypes information in reference \link{Default: Celltypes}}
}
\value{
query seurat object with predictions from ELeFHAnt, Label Transfer and scPred added to meta.data of the object
}
\description{
Celltype annotation is a function to annotate celltypes in a single cell datasets.
Benchmark ELeFHAnt is an easy to use function if the user wants to compare the predictions from ELeFHAnt to Label Transfer and scPred
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
