% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_functions.R
\name{CelltypeAnnotation}
\alias{CelltypeAnnotation}
\title{ELeFHAnt Celltype Annotation}
\usage{
CelltypeAnnotation(
  reference = NULL,
  query = NULL,
  downsample = FALSE,
  downsample_to = 500,
  classification.method = c("randomForest", "SVM", "Ensemble"),
  crossvalidationSVM = 5,
  validatePredictions = TRUE,
  selectvarfeatures = 2000,
  ntree = 500,
  cost = 1,
  classification.approach = "ClassifyCells",
  species = NULL,
  tissue = NULL
)
}
\arguments{
\item{reference}{a processed Seurat Object with Celltypes column in metadata}

\item{query}{a processed seurat object with seurat_clusters column in metadata}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample reference, enabling fast computation. if classification.approach is set to "ClassifyCells_usingApproximation" query will be downsampled along with reference.}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 500}}

\item{classification.method}{choose classification method for learning and predicting celltypes.
Choices: randomForest (decision trees), SVM (Support Vector Machines) or Ensemble (uses estimation robustness of both randomForest and SVM to predict)(}

\item{crossvalidationSVM}{if a integer value k>0 is specified, a k-fold cross validation on the training data is performed to assess the quality of the model (Default: 5)}

\item{validatePredictions}{logical indicator (TRUE or FALSE) to asses predictions by deploying Gene set enrichment analysis and assessing markers from CellMarker Database.}

\item{selectvarfeatures}{number of variable features to select while training (default: 2000)}

\item{ntree}{number of trees randomForest classifier should build (Default: 500)}

\item{cost}{cost of constraints violation (Default: 1)}

\item{classification.approach}{apprach to classify cells 1) ClassifyCells 2) ClassifyCells_usingApproximation. Default: ClassifyCells. We recommend using ClassifyCells_usingApproximation
when reference has significantly less number of cells compared to query}

\item{species}{human or mouse if validatePredictions = TRUE}

\item{tissue}{please check human_tissues or mouse_tissues if validatePredictions = TRUE}
}
\value{
query seurat object with predictions added to meta.data of the object
}
\description{
Celltype annotation is a function to annotate celltypes in a single cell datasets.
It requires a reference dataset (a processed Seurat Object with Celltypes column in metadata) and a query dataset (a processed
seurat object with seurat_clusters column in metadata). One can choose from randomForest, SVM or Ensemble classifiction method
to learn celltypes from reference dataset and then predict celltypes for query dataset.
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
