% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_functions.R
\name{ClassifyCells}
\alias{ClassifyCells}
\title{ELeFHAnt Cell type Annotation per cell in query}
\usage{
ClassifyCells(
  reference = reference,
  query = query,
  downsample = downsample,
  downsample_to = downsample_to,
  classification.method = classification.method,
  crossvalidationSVM = crossvalidationSVM,
  validatePredictions = validatePredictions,
  selectvarfeatures = selectvarfeatures,
  ntree = ntree,
  cost = cost,
  species = NULL,
  tissue = NULL
)
}
\arguments{
\item{reference}{a processed Seurat Object with Celltypes column in metadata}

\item{query}{a processed seurat object with seurat_clusters column in metadata}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample reference, enabling fast computation}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 500}}

\item{classification.method}{choose classification method for learning and predicting celltypes.
Choices: randomForest (decision trees), SVM (Support Vector Machines) or Ensemble (uses estimation robustness of both randomForest and SVM to predict)}

\item{crossvalidationSVM}{if a integer value k>0 is specified, a k-fold cross validation on the training data is performed to assess the quality of the model (Default: 5)}

\item{validatePredictions}{logical indicator (TRUE or FALSE) to asses predictions by deploying Gene set enrichment analysis}

\item{selectvarfeatures}{number of variable features to select while training (Default: 2000)}

\item{ntree}{number of trees randomForest classifier should build (Default: 500)}

\item{cost}{cost of constraints violation (Default: 1)}

\item{species}{human or mouse if validatePredictions = TRUE}

\item{tissue}{please check human_tissues or mouse_tissues if validatePredictions = TRUE}
}
\value{
query seurat object with predictions added to meta.data of the object
}
\description{
Classify Cells is a function to annotate cells in a single cell datasets.
It requires a reference dataset (a processed Seurat Object with Celltypes column in metadata) and a query dataset (a processed
seurat object with seurat_clusters column in metadata). One can choose from randomForest, SVM or Ensemble classifiction method
to learn celltypes from reference dataset and then predict celltypes for query dataset.
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
