% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_Functions.R
\name{CrossSpecies_Conversion}
\alias{CrossSpecies_Conversion}
\title{ELeFHAnt CrossSpecies_Conversion}
\usage{
CrossSpecies_Conversion(
  seurat.objects = c(),
  species_from = NULL,
  species_to = NULL,
  format_from = NULL,
  format_to = NULL
)
}
\arguments{
\item{seurat.objects}{one or more seurat objects}

\item{species_from}{current species: human, mouse, rhesus, zebrafish, chicken and rat}

\item{species_to}{convert to: human, mouse, rhesus, zebrafish, chicken and rat}

\item{format_from}{which format genes are present in the object: ensembl or symbol}

\item{format_to}{which format genes should be converted to: ensembl or symbol}
}
\value{
Cross-species converted Seurat Object
}
\description{
CrossSpecies_Conversion is a function to convert gene names of a or more seurat objects from one species to other. For example:
if your data comes from mouse and you want to compare to human then you can use this function to convert. Currently we support following species: human, mouse, rhesus, zebrafish, chicken and rat
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
