% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_Functions.R
\name{DeduceRelationship}
\alias{DeduceRelationship}
\title{ELeFHAnt Deduce Relationship}
\usage{
DeduceRelationship(
  reference1 = NULL,
  reference2 = NULL,
  downsample = TRUE,
  downsample_to = 1000,
  selectvarfeatures = 2000,
  ntree = 500,
  annotationCol_ref1 = "Celltypes",
  annotationCol_ref2 = "Celltypes"
)
}
\arguments{
\item{reference1}{a processed Seurat Object with Celltypes column in metadata}

\item{reference2}{a processed seurat object with Celltypes column in metadata}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample reference1 and reference2, enabling fast computation}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 1000}}

\item{selectvarfeatures}{number of variable features to select while training (Default: 2000)}

\item{ntree}{number of trees randomForest classifier should build (Default: 500)}

\item{annotationCol_ref1}{Name of column in metdata which has Celltypes information \link{Default: Celltypes} in reference1}

\item{annotationCol_ref2}{Name of column in metdata which has Celltypes information \link{Default: Celltypes} in reference2}
}
\value{
ggplot2 heatmap object and heatmap is automatically saved
}
\description{
Deduce Relationship is a function that hels deduce relationships among celltypes/annotationCol between two datasets.
It requires two datasets (both processed Seurat Objects). Obtain relationships among metadata of interest. Function outputs a heatmap with annotation from dataset1 y-axis and annotation from dataset2 on x-axis.
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
