% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_functions.R
\docType{data}
\name{human_tissues}
\alias{human_tissues}
\title{ELeFHAnt human tisssues
human tissues}
\format{
An object of class \code{character} of length 99.
}
\usage{
human_tissues
}
\description{
ELeFHAnt human tisssues
human tissues
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
\keyword{datasets}
