% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_Functions.R
\name{LabelHarmonization}
\alias{LabelHarmonization}
\title{ELeFHAnt Label Harmonization}
\usage{
LabelHarmonization(
  seurat.objects = c(),
  perform_integration = FALSE,
  integrated.atlas = NULL,
  downsample = TRUE,
  downsample_to = 1000,
  npcs = 30,
  resolution = 0.8,
  validatePredictions = TRUE,
  selectanchorfeatures = 2000,
  ntree = 500,
  k.anchor = 5,
  k.filter = 200,
  k.score = 30,
  dims = 1:30,
  species = NULL,
  tissue = NULL,
  annotationCol = "Celltypes"
)
}
\arguments{
\item{seurat.objects}{a list of processed seurat objects (please set Default Assay to "RNA") with Celltypes column in their respective meta.data to perform integration on}

\item{perform_integration}{logical Indicator (TRUE or FALSE) to perform integration using list of seurat.objects}

\item{integrated.atlas}{an integrated seurat object with CellTypes and seurat_clusters column in meta.data. Required if perform_integration = FALSE}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample Seurat objects or integrated seurat object, enabling fast computation}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 1000}}

\item{npcs}{number of principal components to compute after integration \link{Default: 30}}

\item{resolution}{value of the resolution parameter, decides size of cell communities. \link{Default: 0.8}}

\item{validatePredictions}{logical indicator (TRUE or FALSE) to asses predictions by deploying Gene set enrichment analysis}

\item{selectanchorfeatures}{number of anchor features to use for integrating datasets (Default: 2000)}

\item{ntree}{number of trees randomForest classifier should build (Default: 500)}

\item{k.anchor}{How many neighbors (k) to use when picking anchors \link{Default: 5}}

\item{k.filter}{How many neighbors (k) to use when filtering anchors \link{Default: 200}}

\item{k.score}{How many neighbors (k) to use when scoring anchors \link{Default: 30}}

\item{dims}{Which dimensions to use from the CCA to specify the neighbor search space \link{Default: 1:30}}

\item{species}{human or mouse if validatePredictions = TRUE}

\item{tissue}{please check human_tissues or mouse_tissues if validatePredictions = TRUE}

\item{annotationCol}{Name of column in metdata which has Celltypes information \link{Default: Celltypes}. It has to be same across atlases you want to integrate or have integrated}
}
\value{
integrated seurat object with harmonized celltypes added to meta.data of the object
}
\description{
Label Harmonization is a function to harmonize cell labels (celltypes) across single cell datasets.
It requires a list of processed Seurat Objects or a integrated seurat object.
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
