% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_Functions.R
\docType{data}
\name{mouse_tissues}
\alias{mouse_tissues}
\title{ELeFHAnt mouse tissues}
\format{
An object of class \code{character} of length 293.
}
\usage{
mouse_tissues
}
\description{
Select tissues for CellMarker v2.0 based validation
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
\keyword{datasets}
