% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_functions.R
\name{ValidatePredictions}
\alias{ValidatePredictions}
\title{ELeFHAnt Validate Predictions}
\usage{
ValidatePredictions(
  species = NULL,
  tissue = NULL,
  query = NULL,
  reference = NULL
)
}
\arguments{
\item{species}{human or mouse to select the C8 hallmark cell type gene sets}

\item{tissue}{please check human_tissues or mouse_tissues if validatePredictions = TRUE}

\item{query}{a processed seurat object with seurat_clusters column in metadata}

\item{reference}{a processed seurat object with Celltypes column in metadata}
}
\value{
GSEA result
}
\description{
Validate Predictions is a function to validate celltype assignments. It uses Gene Set Enrichment Analysis (GSEA) and cell type markers from CellMarker Database to let user assess celltype annotation/harmonization.
It requires species, tissue and query dataset. For GSEA, ELeFHAnt uses C8 Hallmark Gene Sets and from CellMarker Database it uses markers curated based on experiments.
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
