% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_functions.R
\name{BenchmarkELeFHAnt}
\alias{BenchmarkELeFHAnt}
\title{Benchmark ELeFHAnt against scPred and Label Transfer}
\usage{
BenchmarkELeFHAnt(
  reference = reference,
  query = query,
  downsample = TRUE,
  downsample_to = 500,
  crossvalidationSVM = 5,
  selectvarfeatures = 2000,
  ntree = 500,
  cost = 1
)
}
\arguments{
\item{reference}{a processed Seurat Object with Celltypes column in metadata}

\item{query}{a processed seurat object with seurat_clusters column in metadata}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample reference, enabling fast computation.}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 500}}

\item{crossvalidationSVM}{if a integer value k>0 is specified, a k-fold cross validation on the training data is performed to assess the quality of the model (Default: 5)}

\item{selectvarfeatures}{number of variable features to select while training (Default: 2000)}

\item{ntree}{number of trees randomForest classifier should build (Default: 500)}

\item{cost}{cost of constraints violation (Default: 1)}
}
\value{
query seurat object with predictions from ELeFHAnt, Label Transfer and scPred added to meta.data of the object
}
\description{
Celltype annotation is a function to annotate celltypes in a single cell datasets.
Benchmark ELeFHAnt is an easy to use function if the user wants to compare the predictions from ELeFHAnt to Label Transfer and scPred
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
