% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_Functions.R
\name{CelltypeAnnotation}
\alias{CelltypeAnnotation}
\title{ELeFHAnt Celltype Annotation}
\usage{
CelltypeAnnotation(
  reference = NULL,
  query = NULL,
  downsample = TRUE,
  downsample_to = 1000,
  validatePredictions = TRUE,
  selectvarfeatures = 2000,
  ntree = 500,
  species = NULL,
  tissue = NULL,
  annotationCol = "Celltypes"
)
}
\arguments{
\item{reference}{a processed Seurat Object with Celltypes column in metadata}

\item{query}{a processed seurat object with seurat_clusters column in metadata}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample reference, enabling fast computation.}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 1000}}

\item{validatePredictions}{logical indicator (TRUE or FALSE) to asses predictions by deploying Gene set enrichment analysis and assessing markers from CellMarker Database.}

\item{selectvarfeatures}{number of variable features to select while training (default: 2000)}

\item{ntree}{number of trees randomForest classifier should build (Default: 500)}

\item{species}{human or mouse if validatePredictions = TRUE}

\item{tissue}{please check human_tissues or mouse_tissues if validatePredictions = TRUE}

\item{annotationCol}{Name of column in metdata which has Celltypes information in reference \link{Default: Celltypes}}
}
\value{
query seurat object with predictions added to meta.data of the object
}
\description{
Celltype annotation is a function to annotate celltypes in a single cell datasets.
It requires a reference dataset and a query dataset. Ensemble classifiction method is used
to learn celltypes from reference dataset and then predict celltypes for query dataset.
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
