% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_functions.R
\name{DeduceRelationship}
\alias{DeduceRelationship}
\title{ELeFHAnt Deduce Relationship}
\usage{
DeduceRelationship(
  reference1 = NULL,
  reference2 = NULL,
  downsample = FALSE,
  downsample_to = 100,
  classification.method = c("randomForest", "SVM", "Ensemble"),
  crossvalidationSVM = 5,
  selectvarfeatures = 2000,
  ntree = 500
)
}
\arguments{
\item{reference1}{a processed Seurat Object with Celltypes column in metadata}

\item{reference2}{a processed seurat object with Celltypes column in metadata}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample reference1 and reference2, enabling fast computation}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 100} in reference and query for Celltypes and seurat_clusters resspectively}

\item{classification.method}{choose classification method for learning and predicting celltypes.
Choices: randomForest (decision trees), SVM (Support Vector Machines) or Ensemble (uses estimation robustness of both randomForest and SVM to predict)}

\item{crossvalidationSVM}{if a integer value k>0 is specified, a k-fold cross validation on the training data is performed to assess the quality of the model}

\item{selectvarfeatures}{number of variable features to select while training (default: 2000)}

\item{ntree}{number of trees randomForest classifier should build (Default: 500)}
}
\value{
ggplot2 heatmap object and heatmap is automatically saved
}
\description{
Deduce Relationship is a function that hels deduce relationships among celltypes between two datasets.
It requires two datasets (both processed Seurat Objects with Celltypes column in metadata). One can choose from randomForest, SVM or Ensemble classifiction method
to obtain relationships among celltypes. Function outputs a heatmap with celltypes from dataset1 as rows and celltypes from dataset2 as columns.
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
