% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_Functions.R
\name{LabelHarmonization}
\alias{LabelHarmonization}
\title{ELeFHAnt Label Harmonization}
\usage{
LabelHarmonization(
  seurat.objects = c(),
  perform_integration = TRUE,
  integrated.atlas = NULL,
  downsample = TRUE,
  downsample_to = 100,
  npcs = 30,
  resolution = 0.5,
  classification.method = c("randomForest", "SVM", "Ensemble"),
  crossvalidationSVM = 10,
  validatePredictions = TRUE,
  selectanchorfeatures = 2000
)
}
\arguments{
\item{seurat.objects}{a list of processed seurat objects (please set Default Assay to "RNA") with Celltypes column in their respective meta.data to perform integration on}

\item{perform_integration}{logical Indicator (TRUE or FALSE) to perform integration using list of seurat.objects}

\item{integrated.atlas}{an integrated seurat object with CellTypes and seurat_clusters column in meta.data. Required if perform_integration = FALSE}

\item{downsample}{logical Indicator (TRUE or FALSE) to downsample seurat objects enabling fast computation}

\item{downsample_to}{a numerical value > 1 to downsample cells \link{Default: 100}}

\item{npcs}{number of principal components to compute after integration}

\item{resolution}{value of the resolution parameter, decides size of cell communities.}

\item{classification.method}{choose classification method for learning and harmonizing celltypes.
Choices: randomForest (decision trees), SVM (Support Vector Machines) or Ensemble (uses learning robustness of both randomForest and SVM to predict)}

\item{crossvalidationSVM}{if a integer value k>0 is specified, a k-fold cross validation on the training data is performed to assess the quality of the model}

\item{validatePredictions}{logical indicator (TRUE or FALSE) to asses predictions by deploying Gene set enrichment analysis}

\item{selectanchorfeatures}{number of anchor features to use for integrating datasets (Default: 2000)}
}
\value{
integrated seurat object with harmonized celltypes added to meta.data of the object
}
\description{
Label Harmonization is a function to harmonize cell labels (celltypes) across single cell datasets.
It requires a list of processed Seurat Objects with Celltypes column in metadata or a integrated seurat object
(seurat object with Celltypes and seurat_clusters column in metadata). One can choose from randomForest, SVM or Ensemble classifiction method
to harmonize celltypes. Please See: DefaultAssay of each object should be set to "RNA"
}
