% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_functions.R
\docType{data}
\name{mouse_tissues}
\alias{mouse_tissues}
\title{ELeFHAnt mouse tisssues
mouse tissues}
\format{
An object of class \code{character} of length 79.
}
\usage{
mouse_tissues
}
\description{
ELeFHAnt mouse tisssues
mouse tissues
}
\author{
Praneet Chaturvedi & Konrad Thorner
}
\keyword{datasets}
