% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELeFHAnt_Functions.R
\name{ValidatePredictions}
\alias{ValidatePredictions}
\title{ELeFHAnt Validate Predictions}
\usage{
ValidatePredictions(reference = NULL, query = NULL)
}
\arguments{
\item{reference}{a processed Seurat Object with Celltypes column in metadata}

\item{query}{a processed seurat object with seurat_clusters column in metadata}
}
\value{
GSEA result
}
\description{
Validate Predictions is a function to validate celltype assignments. It uses Gene Set Enrichment Analysis (GSEA) to asses the enrichment.
It requires a reference dataset (a processed Seurat Object with Celltypes column in metadata) and a query dataset (a processed
seurat object with seurat_clusters column in metadata).
}
